refactor(callbacks): move all callbacks to `callbacks.py`
